2nd Place/requirements.txt
albumentations
ffmpeg_python
numpy
pandas
scikit_learn
timm
tqdm
